Refactor: exception 패키지 분리
