+++ src/main/java/com/cch/codechallengehub/web/exception/custom/AccessDeniedException.java
@@ -1,5 +1,6 @@
-package com.cch.codechallengehub.web.exception;
+package com.cch.codechallengehub.web.exception.custom;
 
+import com.cch.codechallengehub.web.exception.ExceptionBase;
 import org.springframework.http.HttpStatus;
 
 public class AccessDeniedException extends ExceptionBase {
