+++ src/main/java/com/cch/codechallengehub/web/exception/custom/UnauthorizedException.java
@@ -1,5 +1,6 @@
-package com.cch.codechallengehub.web.exception;
+package com.cch.codechallengehub.web.exception.custom;
 
+import com.cch.codechallengehub.web.exception.ExceptionBase;
 import org.springframework.http.HttpStatus;
 
 public class UnauthorizedException extends ExceptionBase {
